fix: prevent navigation lock on empty note

+++ app/src/main/java/com/example/up/data/repository/WeatherRepositoryImpl.kt
@@ -1,9 +1,9 @@
 package com.example.up.data.repository
 
-import com.example.server.data.remote.OpenMeteoServiceApi
 import com.example.up.common.Resource
-import com.example.up.data.remote.mappers.toDomain
 import com.example.up.common.safeApiCall
+import com.example.up.data.remote.OpenMeteoApi.OpenMeteoServiceApi
+import com.example.up.data.remote.mappers.toDomain
 import com.example.up.domain.model.Weather
 import com.example.up.domain.repository.WeatherRepository
 import kotlinx.coroutines.flow.Flow
@@ -11,10 +11,10 @@ import kotlinx.coroutines.flow.Flow
 class WeatherRepositoryImpl(
     val api: OpenMeteoServiceApi
 ) : WeatherRepository{
-    override suspend fun getCurrentWeather(lat: Float, lon: Float): Flow<Resource<Weather>> = safeApiCall {
+    override suspend fun getCurrentWeather(lat: Double, lon: Double): Flow<Resource<Weather>> = safeApiCall {
         api.getCurrentWeather(lat = lat, lon = lon).toDomain()
     }
-    override suspend fun getAvgWeather(lat: Float, lon: Float, date: String): Flow<Resource<Weather>> = safeApiCall {
+    override suspend fun getAvgWeather(lat: Double, lon: Double, date: String): Flow<Resource<Weather>> = safeApiCall {
         api.getHourlyWeather(
             lat = lat,
             lon = lon,
