+++ app/src/main/java/com/example/up/domain/model/Position.kt
@@ -1,6 +1,6 @@
 package com.example.up.domain.model
 
 data class Position(
-    val lat: Float,
-    val lon: Float,
+    val lat: Double,
+    val lon: Double,
 )
